Update README.md
Hand-Rowing-Recumbent-Trikez
My on going Efforts to design and build a hand rowing recumbent Trike
Spelling is not my strong suit due to dyslexia. So be nice Also I have no time for WOKE anything so keep it to yourself.
As I have looked all over for plans and found near to non for such a Trike I finally resorted to designing my own
To that end here is what I have in mind up to the time of this posting
As I have no welding skills I need to find another method to build with To these ends I am thinkin of using several inner twined layers of PVC pipes with solid incert type joints to get the maximum strength with a degree of flexability to my build. With no welding needed.
I have purchased several secound hand bicycles from which I will reuse some parts. Primaraly Wheel Tires frunt forks and possibly a rear fram portion brakes chains gears and others as devised.
I found a design of a gear and sleave that is as follows The sleave is an elongated elipse with gear teeth at the top and bottom centers along it's parelel longer inner sides there are just enough of these teeth to keep the set of corisponding teeth on the center gear always engaged. with enough further length in its inner cut out for the gears teeth to freely move around just bearly leaving the top set of teeth before instantly being ingaged with the lower set of teeth There by causing the center gear to be in constant motion in the same direction.
The center gear has teeth on about 1/3rd of its cercumfrence. Just enough so as to always be in contact with one or the other sets of teeth on the upper or lower part of the sleave. 
I hope to add a neutral position to this set up by spring loading the sleave to be in the ingaged position by default. And by applying a downward and thus outward motion to the ores as it were The sleave can be disingaged from the Gear temparaly. And via the springs action be reingaged when this pressure is realeased. 
Not sure if this design will allow for a reverse motion in the system or if this may have to be addressed in another way.
I was planning to use faberic to make pour boy fiberglass but after thinking about it I may go with a simple layering of plastic grocerie sack from smith's as I like the way they turn brown when in several layers and they can be added to PVC without any glue ( just heating with a cloths iron )
These layers of plastic will allow me to form smooth flat areas were needed rather than have the coragated effect of several PVC pipes layed up side by side.
I plan on reusing parts off of old bikes where posible to save on costs. Like the front forks to be used to carry the front wheels and head lights. 
also the wheels and tires chains gears frame parts brakes and cables for same.
As this is to be a hand rowing system there will be no need for pedals Nore the bike seat.
I think front forks off two smaller bikes  will work well for the front stearing hubs. These will be mounted up under the fender / frame arms and by using tie rods they can be steard in unisence via a set of foot pegs
Just at the top of these stearing towers will be the head light housing so that the head lights will always be facing in the same direction as the wheels
The tail light assembly will be atop a rear wheel in their own inclosure. In addition I'm thinking of adding what I call a God Light front and back for those special peopple that refuse to reduce thier high beams when approaching. Or coming up behind. I also want to have a headlight above my head for extra night vision
The braking will be barrowed directly from the Bikes and mounted onto the new handels.
As I currently have 2 sets of stunt pegs off the used bikes I will use them mounted on the ends of some all thread rods and atatched to the tops of the front mounting towers to give me foot stearing
I am up in the air as to how to do the seat. On the one hand I want to do a hand braided net type or maybe a rag carpet type then again I might go with a plastic molded version like in race cars.
If my power system will allow for it I want to have a Semi Truck horn for it. 
I'm not sure about any liciensing  on it but as my daughter is starting at the DMV tomarrow I am having her look into it for me.
If and when I get this to a point I am happy with it I will offer a set of plans a kit for this basic modle and or a basic customed out Trickerr Trike design or a fully customer designed custom if posible. ( I say this because just while collecting up the used bikes for this build I found 2 potential customers for them )
And because I can build them cheaper than building with metal. I can sell them for less and still pull a prophet. Oh and have a hella lot of fun doing it.
And in an effort to do this I will be building a Custom Trike Club from amoungst my customers and other builders.
Wait till I get this going good and then see what I'm going to do with these micro mini EV PUs out of China. You see I was into building custom vans back when that was big